Move the shared logic into the base class.
