+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Settings/Language/LanguageIssueSection.cs
@@ -18,7 +18,7 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings.Langua
     {
         protected override LyricEditorMode EditMode => LyricEditorMode.Language;
 
-        protected override LyricsIssueTable CreateIssueTable() => new LanguageIssueTable();
+        protected override LyricsIssueTable CreateLyricsIssueTable() => new LanguageIssueTable();
 
         private partial class LanguageIssueTable : LyricsIssueTable
         {
